Extract behavior capture/compare out of test_util into behavior_compare
test_util.py had grown to conflate two concerns: the pytest harness
(check_states/check_exec_err/check_messages/simplefs) and the machinery
for capturing and comparing execution behavior (flexible_equal,
summarize_execution, compare_results/compare_returns, run_differential,
_is_opaque, ...). The latter is imported by runtime code (diff_behavior,
tools/measure_support), so "test_util" was a misnomer for it.

Move that half into crosshair/behavior_compare.py and repoint all
importers; rename test_util_test.py -> behavior_compare_test.py. The two
halves had no cross-references, so this is a pure relocation with no
behavior change. test_util.py drops from 632 to 110 lines.
