+++ crosshair/libimpl/collectionslib_ch_test.py
@@ -3,9 +3,9 @@ from typing import DefaultDict, Deque, Optional, Sequence
 
 import pytest  # type: ignore
 
+from crosshair.behavior_compare import ResultComparison, compare_results
 from crosshair.core_and_libs import MessageType, analyze_function, run_checkables
 from crosshair.options import AnalysisOptionSet
-from crosshair.test_util import ResultComparison, compare_results
 
 # deque
