+++ crosshair/libimpl/decimallib_ch_test.py
@@ -5,8 +5,12 @@ from typing import Union
 
 import pytest  # type: ignore
 
+from crosshair.behavior_compare import (
+    ResultComparison,
+    compare_results,
+    compare_returns,
+)
 from crosshair.core_and_libs import MessageType, analyze_function, run_checkables
-from crosshair.test_util import ResultComparison, compare_results, compare_returns
 
 
 def _binary_op_under_context(ctx, op):
