+++ crosshair/libimpl/jsonlib_ch_test.py
@@ -4,9 +4,9 @@ from typing import Dict, List, Tuple, Union
 
 import pytest  # type: ignore
 
+from crosshair.behavior_compare import compare_results
 from crosshair.core_and_libs import MessageType, analyze_function, run_checkables
 from crosshair.options import AnalysisOptionSet
-from crosshair.test_util import compare_results
 
 
 def check_decode(s: str):
